Added PuckWorld example

+++ src/ConvNet/Utilities.h
@@ -82,7 +82,7 @@ public:
 	Volume(int, int) {Panic("TODO");}
 	Volume(int width, int height, int depth, Volume& vol);
 	Volume(int width, int height, int depth, VolumeDataBase& weights);
-	Volume(int width, int height, int depth, const Vector<int>& weights);
+	Volume(int width, int height, int depth, const Vector<double>& weights);
 	Volume(const Volume& o) {owned_weights = false; weights = NULL; *this = o;}
 	Volume(int width, int height, int depth); // Volume will be filled with random numbers
 	Volume(int width, int height, int depth, double default_value);
